making changes according to detekt
