+++ app/src/main/java/com/karthik/splash/homescreen/bottomhometab/HomeViewModel.kt
@@ -8,7 +8,8 @@ class HomeViewModel: ViewModel() {
     private val _selectedCategory: MutableLiveData<BottomHomeTabType> = MutableLiveData()
     val selectedCategory: LiveData<BottomHomeTabType> = _selectedCategory
 
-    fun initalSelectedCategory(category: BottomHomeTabType) = onCategorySelected(category = category)
+    fun initalSelectedCategory(category: BottomHomeTabType) =
+        onCategorySelected(category = category)
 
     fun onCategorySelected(category: BottomHomeTabType) {
         _selectedCategory.postValue(category)
